backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
anthropic
openai
pandas
pydantic